Administrator feature

+++ DFS/src/mainwidget.cpp
@@ -2,8 +2,10 @@
 #include <QMessageBox>
 #include <QVBoxLayout>
 #include <QHeaderView>
+#include <iostream>
 
 #include "mainwidget.hpp"
+#include "clientmacros.hpp"
 
 namespace SHIZ{
 	MainWidget::MainWidget(Logger* logger, NetworkManager* manager, QWidget* parent)
@@ -21,8 +23,8 @@ namespace SHIZ{
 		connect(filterLineEdit, &QLineEdit::textChanged, this, &MainWidget::onFilterTextChanged);
 
 		fileTableWidget = new QTableWidget(this);
-		fileTableWidget->setColumnCount(4);
-		fileTableWidget->setHorizontalHeaderLabels({"File Name", "Owner", "Size (bytes)", "Upload Date"});
+        fileTableWidget->setColumnCount(5);
+        fileTableWidget->setHorizontalHeaderLabels({"File Name", "Owner", "Size (bytes)", "Upload Date", "Group"});
 		fileTableWidget->horizontalHeader()->setSectionResizeMode(QHeaderView::Stretch);
 		fileTableWidget->setSelectionBehavior(QAbstractItemView::SelectRows);
 		fileTableWidget->setSortingEnabled(true);
@@ -57,36 +59,54 @@ namespace SHIZ{
 
 
 	void MainWidget::onDeleteButtonClicked() {
+        onRefreshButtonClicked();
+        setRights();
+
 		int selectedRow = fileTableWidget->currentRow();
 		if (selectedRow >= 0) {
-			QString fileName = fileTableWidget->item(selectedRow, 0)->text();
-			bool success = networkManager->deleteFile(fileName);
-			if (success) {
-				QMessageBox::information(this, "Delete", "File deleted successfully.");
-				onRefreshButtonClicked();
-			} else {
-				QMessageBox::warning(this, "Delete", "File deletion failed.");
-			}
+            QString fileName = fileTableWidget->item(selectedRow, 0)->text();
+
+            if(ACCESS_DELETE){
+                bool success = networkManager->deleteFile(fileName);
+                if (success) {
+                    QMessageBox::information(this, "Delete", "File deleted successfully.");
+                    onRefreshButtonClicked();
+                } else {
+                    QMessageBox::warning(this, "Delete", "File deletion failed.");
+                }
+            }else{
+                QMessageBox::warning(this, "Delete", "You don't have right delete this file.");
+            }
+
 		} else {
 			QMessageBox::warning(this, "Delete", "No file selected.");
 		}
 	}
 
 	void MainWidget::onDownloadButtonClicked(){
-		int selectedRow = fileTableWidget->currentRow();
-		if (selectedRow >= 0) {
-			QString fileName = fileTableWidget->item(selectedRow, 0)->text();
-			QString directory = QFileDialog::getExistingDirectory(this, "Select Download Folder");
+        onRefreshButtonClicked();
+        setRights();
 
-			if (!directory.isEmpty()) {
-				QString filePath = directory + "/" + fileName;
-				bool success = networkManager->downloadFile(filePath);
-				if (success) {
-					QMessageBox::information(this, "Download", "File downloaded successfully.");
-				} else {
-					QMessageBox::warning(this, "Download", "File download failed.");
-				}
-			}
+		int selectedRow = fileTableWidget->currentRow();
+
+		if (selectedRow >= 0) {
+            QString fileName = fileTableWidget->item(selectedRow, 0)->text();
+
+            if(ACCESS_READ){
+                QString directory = QFileDialog::getExistingDirectory(this, "Select Download Folder");
+
+                if (!directory.isEmpty()) {
+                    QString filePath = directory + "/" + fileName;
+                    bool success = networkManager->downloadFile(filePath);
+                    if (success) {
+                        QMessageBox::information(this, "Download", "File downloaded successfully.");
+                    } else {
+                        QMessageBox::warning(this, "Download", "File download failed.");
+                    }
+                }
+            }else{
+                QMessageBox::warning(this, "Download", "You don't have right download this file.");
+            }
 		} else {
 			QMessageBox::warning(this, "Download", "No file selected.");
 		}
@@ -106,21 +126,27 @@ namespace SHIZ{
 	}
 
 	void MainWidget::onRefreshButtonClicked(){
-		QStringList files = networkManager->requestFileList();
+        setRights();
+        QStringList files = networkManager->requestFileList(currentLogin);
 
+        if(files.size() > 0 && files[0] == RESPONSE_USER_DOES_NOT_EXIST){
+            emit userBanned();
+        }
 		fileTableWidget->setRowCount(files.size());
 		for (int i = 0; i < files.size(); ++i) {
 			QStringList fileInfo = files[i].split("|");
-			if (fileInfo.size() == 4) {
+            if (fileInfo.size() == 5) {
 				QTableWidgetItem* fileNameItem = new QTableWidgetItem(fileInfo[0]);
 				QTableWidgetItem* ownerItem = new QTableWidgetItem(fileInfo[1]);
 				QTableWidgetItem* sizeItem = new QTableWidgetItem(fileInfo[2]);
-				QTableWidgetItem* dateItem = new QTableWidgetItem(fileInfo[3]);
+                QTableWidgetItem* dateItem = new QTableWidgetItem(fileInfo[3]);
+                QTableWidgetItem* groupItem = new QTableWidgetItem(fileInfo[4]);
 
 				fileTableWidget->setItem(i, 0, fileNameItem);
 				fileTableWidget->setItem(i, 1, ownerItem);
 				fileTableWidget->setItem(i, 2, sizeItem);
 				fileTableWidget->setItem(i, 3, dateItem);
+                fileTableWidget->setItem(i, 4, groupItem);
 			}
 		}
 	}
@@ -130,6 +156,13 @@ namespace SHIZ{
 	}
 
 	void MainWidget::onUploadButtonClicked(){
+        onRefreshButtonClicked();
+        setRights();
+
+        if(!ACCESS_WRITE){
+            QMessageBox::warning(this, "Upload", "You don't have right upload files.");
+            return;
+        }
 		QString filePath = QFileDialog::getOpenFileName(this, "Select a file to upload");
 		if (filePath.isEmpty()) return;
 		QString fileName = QFileInfo(filePath).fileName();
@@ -160,4 +193,30 @@ namespace SHIZ{
 			QMessageBox::warning(this, "Upload", "File upload failed.");
 		}
 	}
+
+    void MainWidget::setRights(){
+        QString rights = networkManager->getUserInfo(currentLogin).split("|")[0];
+
+        QChar right_to_read('r');
+        QChar right_to_write('w');
+        QChar right_to_delete('d');
+
+        if(rights.contains(right_to_read)){
+            ACCESS_READ = true;
+        }else{
+             ACCESS_READ = false;
+        }
+
+        if(rights.contains(right_to_write)){
+            ACCESS_WRITE = true;
+        }else{
+            ACCESS_WRITE = false;
+        }
+
+        if(rights.contains(right_to_delete)){
+            ACCESS_DELETE = true;
+        }else{
+            ACCESS_DELETE = false;
+        }
+    }
 }
